Move to `logos` subfolder
