+++ packages/components/src/logos/wordpress-logo/README.md
