+++ packages/components/src/logos/wordpress-logo/index.tsx
