+++ packages/components/src/logos/wordpress-logo/style.scss
